% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_all.R
\name{Compare_meta}
\alias{Compare_meta}
\title{Compare_meta}
\usage{
Compare_meta(dat, config, prep = 0.2, method = "all", C = NULL, ...)
}
\arguments{
\item{dat, }{matrix or data.frame, row is sample id}

\item{config}{, data.frame. group information}

\item{prep}{, default = 0.2,cutoff of filter}

\item{method}{, character, "all" or specific method}

\item{C}{, confounders only available for locom or ancombc}

\item{...}{}
}
\value{
list
}
\description{
Compare_meta
}
