% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sim_diri.R
\name{Sim_diri}
\alias{Sim_diri}
\title{Sim_diri}
\usage{
Sim_diri(
  pi = NULL,
  n,
  disp = 0.02,
  beta = 3,
  sigN = NULL,
  have.bias = FALSE,
  disp2 = 0.2
)
}
\arguments{
\item{pi, }{vector: the parameter of dirichlet distirbution}

\item{n, }{numeric , sample number}

\item{disp}{, numeric , shape of the simulation}

\item{beta}{, numeric , generate the causal taxa}

\item{sigN}{, numeric, number of causal taxa}

\item{have.bias}{, logistic, if add a bias factor to simualted data}

\item{disp2}{, numeric, the other parameter of beta distribution}
}
\value{
matrix or list
}
\description{
Sim_diri
}
